Rename to upload-utils.server.ts
